+++ player/states/PlayerState.gd
@@ -1,0 +1,5 @@
+extends State
+
+class_name PlayerState
+
+onready var player: Player = get_tree().get_nodes_in_group("player")[0]